backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==21.2.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.1
Django==3.1
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
djangorestframework==3.11.1
djangotoolbox==1.8.0
drf-spectacular==0.19.0
filelock==3.0.12
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.4
mysqlclient==2.0.1
numpy==1.22.0
pandas==1.3.5
path==13.1.0
Pillow==8.4.0
psycopg2==2.8.5
pycparser==2.21
Pygments==2.6.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2020.1
PyYAML==5.4.1
requests==2.26.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.26.7
zipp==3.1.0
